+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.7.0 2020/11/24
+
+Feature
+- add `account` skill to pick from account explorers (or defi dashboard) list. The default option is randomly open the given address with an account explorer.
+- add Defi Balance skills `debank`, `zapper`, `zerion`
+Enhancement
+- Applied more emojis
+
 # 0.6.1 2020/11/23
 
 Feature